(1/2) Fix multiple definitions
By default, eLua instructs the linker to accept multiple symbol
definitions. This means that if a symbol named `A` is defined in two
objects `O1` and `O2` the linker will use both of these definitions
instead of complaining that `A` is defined multiple times (which is the
default linker behaviour). This is problematic for a number of reasons:

- If `A` is a funtion, the programmer will have no way to know that some
of the functions that he wrote might be defined somewhere else. This
might lead to an incorrect program behaviour.
- If `A` is a variable, multiple copies of `A` will end up in different
object files, which wastes memory and might also lead to an incorrect
program behaviour.

This PR forbids multiple symbol definitions and fixes the multiple
symbol definitions that were found in the code:

- `port_std` defined in `cpu_sim3u167.h` was instantiated multiple
times, now it is instantiated once.
- `mySystemInit` in `platform.c` wasn't used at all (the implementation
in `src/platform/sim3u1xx/FreakUSB/hw/sim3u1xx/hw.c` is used instead).
- `armed_flash_key` in `platform.c` wasn't used at all, but it was
causing "multiple symbol definition" errors.